[Add] Userのエンドポイントを追加 idをなくしてuuidのみで判定する

+++ app/controllers/api/v1/users_controller.rb
@@ -7,14 +7,18 @@ class Api::V1::UsersController < Api::V1::BaseController
   end
 
   def show
-    user = User.find(params[:id])
+    user = User.find_by(uuid: params[:uuid])
     render json: user
   end
 
   def destroy
-    user = User.find(params[:id])
-    user.destroy!
-    render json: { message: '何を返そうか' }
+    user = User.find_by(uuid: params[:uuid])
+    if user == current_user
+      user.destroy!
+      render json: { message: '何を返そうか' }
+    else
+      head 404
+    end
   end
 
   def current
